Retain output files between runs
Previously, these scripts would clean up output files as soon as the
scripts were done running. This made it impossible to review the
output files and troubleshoot in respons to queries from NCBI.

This commit cleans up output files at the beginning of the process,
allowing output to remain in place between subsequent runs of the
scripts.